requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-crispy-forms==1.11.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
idna==2.10
isodate==0.6.0
Pillow==8.1.2
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
